refactor(save-link): remove legacy DDB writers superseded by the Article aggregate
The Article aggregate is now the sole writer of crawlStatus and summaryStatus
(PRs #292, #293, #294, #297). The legacy dynamodb-article-crawl and
dynamodb-generated-summary files still owned the transient crawlStage /
summaryStage progress markers, so extract those into focused single-purpose
files (mark-crawl-stage / mark-summary-stage), repoint the consumers, and
delete the legacy code together with the now-orphaned mark-crawl-{ready,
failed,unsupported} and saveGeneratedSummary / mark-summary-{failed,skipped,
pending} / findGeneratedSummary exports.

Stage writes deliberately bypass the aggregate: every transition would
otherwise have to thread the current stage, which it shouldn't. The new
files preserve the same unconditional 'SET <stage>' UpdateExpression as
before; coverage is 100% on both happy path and DDB error rethrow.
