+++ projects/save-link/src/generate-summary/link-summariser.ts
@@ -4,7 +4,7 @@ import { z } from "zod";
 import type { SummarySkipReason } from "@packages/article-state-types";
 import type { HutchLogger } from "@packages/hutch-logger";
 import type { CreateAiMessage } from "./create-ai-message.types";
-import type { MarkSummaryStage } from "./article-summary.types";
+import type { MarkSummaryStage } from "./mark-summary-stage";
 import { MAX_EXCERPT_LENGTH, MAX_SUMMARY_LENGTH } from "./max-summary-length";
 
 const SUMMARIZE_PROMPT = readFileSync(
